Add a form that lets you enter a URL, and parses it.
Sorta WIP in that the form will always print out an error, but if
the question URL can be parsed then it prints out the site name and
question ID.

Introduce a SiteList object that can hold the list of sites indexed
in a couple of useful ways; have the Sites actor return it.

Add i18n support (not 100% universal but it's used in the form).

Add Twirl helpers to generate some Bootstrap parts, including the
core page layout.