Remember things we want to or have fetched.
This adds database storage for abstract fetch records, kinds of posts
(questions), and kinds of fetches (metadata, bodies).  There is enough
wiring here so that, when you request a question URL, it looks up in
the database whether or not it needs to be fetched, and records it if
it's not yet known; but there is no actual API call to do the fetch
right now.